feat(#3): svg logo (#6)
* feat: add logo to svg sprite

* feat: update css

* chore: increase version number

* feat: replace text with svg

in both the header and body

* fix: remove unneeded css

* fix: update logo css

* chore: clean up css

when i was working on this earlier this css neatly displayed the logo in the correct size. somehow after a change it no longer does. i’ve inlined the sizes (good for paints anyway) and it works correctly. so this css is no longer needed.

* chore: remove unused css

there’s no text left inside a.logo, plus the `text-decoration` is handled in logo.css